Added AlphaBeta with new Eval
Semi-decent eval that makes assassins retarded

+++ IIA-TP1/Assets/Scripts/UtilityFunction.cs
@@ -8,9 +8,9 @@ public class UtilityFunction
     public float evaluate(State s)
     {
 		if(s.PlayersUnits.Count == 0)
-            return Int32.MinValue;
+            return float.MinValue;
         else if(s.AdversaryUnits.Count == 0)
-            return Int32.MaxValue;
+            return float.MaxValue;
         return 0;
     }
 }
